requirements.txt

# Runtime dependencies for Local Agent Embedded

#Local LLM
llama-cpp-python==0.3.15

# Fuzzy inference
rapidfuzz==3.13.0
tbb>=2021.6.0

# Audio capture / processing
numpy==2.2.6
pyaudio==0.2.14
webrtcvad==2.0.10
pyyaml==6.0.2
typeguard==4.4.4

#text-to-speech
piper-tts==1.3.0
torch==2.8.0

# Speech-to-text
torchaudio==2.8.0
onnx==1.18.0
onnxruntime==1.22.1
setuptools-rust
openai-whisper

#Wake Word
vosk==0.3.45


